pyproject.toml
[tool.poetry]
name = "urbaflow"
version = "0.1.0"
description = "Intégration des données liées à l'urbanisme"
authors = ["Thomas Brosset <thomas.brosset@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
GeoAlchemy2 = "^0.13.3"
simplejson = "^3.19.1"
pytz = "^2023.3"
prefect = "^1.4"
coverage = "^7.2.7"
pytest = "^7.3.1"
geopandas = "^0.13.2"
SQLAlchemy = "^2.0.16"
psycopg2-binary = "^2.9.6"
typer = "^0.9.0"
duckdb = "^0.8.1"
chardet = "^5.2.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
